requirements.txt
pydub
fleep
google-cloud-speech
google-cloud-storage
rootpath
python-dotenv
pynput
pyaudio
sphinx
